+++ ChangeLog.md
@@ -15,6 +15,7 @@ and this project adheres to
 * The ``neighbors`` argument to ``env_neighbors`` for ``EnvironmentMotifMatch`` class.
 * The ``neighbors`` argument to ``cluster_neighbors`` for ``EnvironmentCluster`` class.
 * `freud.order.Nematic` uses orientation vectors instead of quaternions and a nematic director.
+* `freud.order.Nematic` raises a warning when the zero is vector passed.
 * Remove zero-padding from arrays in `freud.environment.EnvironmentCluster` and `freud.environment.EnvironmentMotifMatch` and replace with ragged lists of NumPy arrays.
 
 ### Removed